# crimson/.../ephemeral: faster seastore eph dev