@chanyoung 2025-08-10
jenkins test make check arm64